+++ src/flitr/modules/flitr_image_processors/photometric_equalise/fip_photometric_equalise.cpp
@@ -101,7 +101,7 @@ bool FIPPhotometricEqualise::trigger()
             const size_t componentsPerLine=componentsPerPixel * width;
             const size_t componentsPerImage=componentsPerLine * height;
             
-            size_t y=0;
+            int y=0;
             {
 #ifdef _OPENMP
                 //omp_set_num_threads(4);//There might be some blocking operation or mem bandwidth limited code??? because the parallel for seems to work best with much more threads than CPU cores.
@@ -114,7 +114,7 @@ bool FIPPhotometricEqualise::trigger()
 #ifdef _OPENMP
 #pragma omp for nowait
 #endif
-                    for (y=0; y<height; y++)
+                    for (y=0; y<(int)height; y++)
                     {
                         const size_t lineOffset=y * componentsPerLine;
                         
@@ -132,7 +132,7 @@ bool FIPPhotometricEqualise::trigger()
             
             //Single thread!
             uint32_t imageSum=0;
-            for (y=height/4; y<(height*3)/4; y++)
+            for (y=(int)height/4; y<((int)height*3)/4; y++)
                 //for (y=0; y<height; y++)
             {
                 imageSum+=lineSumArray_[y];
@@ -155,7 +155,7 @@ bool FIPPhotometricEqualise::trigger()
 #ifdef _OPENMP
 #pragma omp for nowait
 #endif
-                    for (y=0; y<height; y++)
+                    for (y=0; y<(int)height; y++)
                     {
                         const size_t lineOffset=y * componentsPerLine;
